LibGL+LibGPU+LibSoftGPU: Move ImageFormat.h to LibGPU
